+++ webviz_subsurface/plugins/_well_completions/__init__.py
@@ -1,0 +1,1 @@
+from ._plugin import WellCompletions
